+++ weblate/trans/models/component.py
@@ -516,11 +516,13 @@ class Component(models.Model, PathMixin, CacheKeyMixin, ComponentCategoryMixin):
             "will cause automatic translation in this one"
         ),
     )
+    # This should match definition in WorkflowSetting
     enable_suggestions = models.BooleanField(
         verbose_name=gettext_lazy("Turn on suggestions"),
         default=True,
         help_text=gettext_lazy("Whether to allow translation suggestions at all."),
     )
+    # This should match definition in WorkflowSetting
     suggestion_voting = models.BooleanField(
         verbose_name=gettext_lazy("Suggestion voting"),
         default=False,
@@ -528,6 +530,7 @@ class Component(models.Model, PathMixin, CacheKeyMixin, ComponentCategoryMixin):
             "Users can only vote for suggestions and can’t make direct translations."
         ),
     )
+    # This should match definition in WorkflowSetting
     suggestion_autoaccept = models.PositiveSmallIntegerField(
         verbose_name=gettext_lazy("Autoaccept suggestions"),
         default=0,
@@ -2483,10 +2486,6 @@ class Component(models.Model, PathMixin, CacheKeyMixin, ComponentCategoryMixin):
             from weblate.trans.tasks import cleanup_component
 
             transaction.on_commit(lambda: cleanup_component.delay(self.id))
-
-        # Send notifications on new string
-        for translation in translations.values():
-            translation.notify_new(request)
 
         if was_change:
             if self.needs_variants_update: